Move node styling types and methods to a separate file
